+++ scripts/bases-constants.mjs
@@ -1,5 +1,4 @@
 export default class Constants {
 	static MODULE_ID = 'bases';
 	static MODULE_NAME = "Bugbear's Assign Status Effects Sorter";
-	static STYLE_ID = 'bases-custom-hud-style';
 }